wfrun2vue - refactor some wf run data parsing out of Backbone into more reusable class.
